requirements.txt
joblib==1.1.0
numpy==1.23.1
opencv_python==4.7.0.68
pandas==1.3.4
scikit_learn==1.1.1
scipy==1.8.1
torch_geometric==2.2.0
tqdm==4.64.0
networkx==3.2.1
pyyaml==6.0.2
pytorch_metric_learning==2.8.1